requirements.txt
absl-py==0.11.0
aiohttp==3.7.4.post0
alabaster==0.7.12
algopy==0.5.7
anyio==3.6.2
appdirs==1.4.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.3.0
asteval==0.9.27
asttokens==2.0.8
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
batchgenerators==0.24
bcrypt==3.1.7
bitstring==3.1.6
bleach==6.0.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
click==7.1.1
cloudpickle==1.6.0
colorlog==4.6.2
configparser==5.0.2
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.16
dataclasses==0.6
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.11
dicom2nifti==2.3.0
distlib==0.3.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.16
ecdsa==0.15
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.3
filelock==3.0.12
fire==0.4.0
fonttools==4.34.4
fpdf2==2.5.5
fsspec==2021.7.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.24.0
google-auth-oauthlib==0.4.2
GPUtil==1.4.0
grpcio==1.34.0
hiddenlayer==0.2
humanize==3.2.0
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==1.6.0
importlib-resources==5.1.3
ipaddress==1.0.23
jedi==0.18.1
Jinja2==2.11.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-server==1.23.4
jupyterlab-pygments==0.2.2
kiwisolver==1.3.1
liac-arff==2.4.0
linecache2==1.0.0
lmfit==1.0.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
MedPy==0.4.0
mistune==0.8.4
mock==4.0.1
more-itertools==8.2.0
mpmath==1.2.1
multidict==5.1.0
nbclassic==1.0.0
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.8.0
nest-asyncio==1.5.5
netaddr==0.7.19
netifaces==0.11.0
networkx==2.5
nibabel==3.2.1
nose==1.3.7
notebook==6.5.4
notebook_shim==0.2.3
numdifftools==0.9.40
numpy==1.23.3
oauthlib==3.1.0
opencv-python==4.5.2.52
packaging==21.3
pandas==1.2.0
pandocfilters==1.5.0
paramiko==2.7.1
parso==0.8.3
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.2
paycheck==1.0.2
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pkgutil_resolve_name==1.3.10
pluggy==0.13.1
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.14.0
psutil==5.7.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.8.1
py-expression-eval==0.3.9
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.17
pyDeprecate==0.3.1
pydicom==2.1.2
Pygments==2.5.2
PyNaCl==1.3.0
pynvml==11.5.0
pyparsing==2.4.7
pyrsistent==0.19.3
pytest==5.4.1
python-dateutil==2.8.2
python-gdcm==3.0.14
pytorch-lightning==1.7.6
pytz==2020.5
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==24.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
rt-utils==1.1.3
scandir==1.10.0
scikit-image==0.18.1
scikit-learn==1.1.1
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.8.2
sentry-sdk==1.0.0
setproctitle==1.3.2
setuptools-scm==3.5.0
shortuuid==1.0.1
SimpleITK==2.1.0
six==1.15.0
sklearn==0.0
smmap==4.0.0
sniffio==1.3.0
snowballstemmer==2.0.0
soupsieve==2.3.2.post1
Sphinx==3.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.1
stack-data==0.5.0
statannotations==0.4.4
statsmodels==0.13.2
subprocess32==3.5.4
Surface-Distance-Based-Measures==0.1
sympy==1.11
tabulate==0.8.7
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2021.1.11
toml==0.10.0
torch==1.12.1
torchio==0.18.84
torchmetrics==0.9.3
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traceback2==1.4.0
traitlets==5.4.0
typing_extensions==4.3.0
ujson==2.0.3
uncertainties==3.1.7
unittest2==1.1.0
urllib3==1.26.2
virtualenv==20.0.18
wandb==0.13.3
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==1.0.1
widgetsnbextension==3.6.1
wrapt==1.12.1
xlrd==1.2.0
yarl==1.6.3
zipp==1.2.0
